# public/openapi.yaml
openapi: 3.1.0
info:
  title: Lune API
  version: 0.3.0
  description: >
    API publique (auth, santé) et API app (PERSONAL / PRO). Toutes les IDs BigInt sont sérialisées en string.
    Toutes les réponses mutables exigent CSRF (Origin) et renvoient `x-request-id` + `cache-control: no-store`.
servers:
  - url: http://localhost:3000
    description: Développement local
  - url: https://diwanbg.work
    description: Production (Railway)
security:
  - cookieAuth: []
...
paths:
...
  /api/pro/businesses/{businessId}/settings:
    get:
      tags: [Pro / Settings]
      security: [ { cookieAuth: [] } ]
      summary: Récupère les paramètres d’entreprise
      description: >
        Crée des paramètres par défaut si absents. Requiert rôle VIEWER+.
        Toutes les réponses incluent `x-request-id` et `cache-control: no-store`.
      parameters: [ { $ref: '#/components/parameters/BusinessId' } ]
      responses:
        '200':
          description: Paramètres
          content:
            application/json:
              schema:
                type: object
                properties:
                  item: { $ref: '#/components/schemas/BusinessSettings' }
    patch:
      tags: [Pro / Settings]
      security: [ { cookieAuth: [] } ]
      summary: Met à jour les paramètres d’entreprise
      description: ADMIN/OWNER uniquement. CSRF (Origin) requis + rate limit.
      parameters: [ { $ref: '#/components/parameters/BusinessId' } ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                invoicePrefix: { type: string }
                quotePrefix: { type: string }
                defaultDepositPercent: { type: integer, minimum: 0, maximum: 100 }
                paymentTermsDays: { type: integer, minimum: 0, maximum: 365 }
                enableAutoNumbering: { type: boolean }
                vatRatePercent: { type: integer, minimum: 0, maximum: 100 }
                vatEnabled: { type: boolean }
                allowMembersInvite: { type: boolean }
                allowViewerExport: { type: boolean }
                integrationStripeEnabled: { type: boolean }
                integrationStripePublicKey: { type: string, nullable: true }
      responses:
        '200':
          description: Paramètres mis à jour
          content:
            application/json:
              schema:
                type: object
                properties:
                  item: { $ref: '#/components/schemas/BusinessSettings' }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: auth_token
  parameters:
    BusinessId:
      name: businessId
      in: path
      required: true
      schema: { type: string }
...
  schemas:
...
    BusinessSettings:
      type: object
      properties:
        id: { type: string }
        businessId: { type: string }
        invoicePrefix: { type: string }
        quotePrefix: { type: string }
        defaultDepositPercent: { type: integer }
        paymentTermsDays: { type: integer }
        enableAutoNumbering: { type: boolean }
        vatRatePercent: { type: integer }
        vatEnabled: { type: boolean }
        allowMembersInvite: { type: boolean }
        allowViewerExport: { type: boolean }
        integrationStripeEnabled: { type: boolean }
        integrationStripePublicKey: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...